Add detect_bottom() dispatcher function for bottom line detection Description:
* Introduced detect_bottom() in mask/api.py to act as a unified interface for bottom detection methods
* Replaced the earlier class-based approach with simpler functional structure under mask/seafloor_detection
* Output is now a 1D DataArray (no channel dim), with one depth per ping
