+++ echopype/mask/__init__.py
@@ -1,3 +1,3 @@
-from .api import apply_mask, frequency_differencing, seafloor_mask_basic
+from .api import apply_mask, detect_seafloor, frequency_differencing
 
-__all__ = ["frequency_differencing", "apply_mask", "seafloor_mask_basic"]
+__all__ = ["frequency_differencing", "apply_mask", "detect_seafloor"]
