+++ echopype/mask/api.py
@@ -2,22 +2,18 @@ import datetime
 import operator as op
 import pathlib
 import sys
-from typing import List, Optional, Union
+from typing import Dict, List, Optional, Union
 
 import dask
 import dask.array
 import numpy as np
 import xarray as xr
 
+# for seafloor detection
+from echopype.mask.seafloor_detection.detect_basic import detect_basic
+
 from ..utils.io import validate_source
 from ..utils.prov import add_processing_level, echopype_prov_attrs, insert_input_processing_level
-
-# for seafloor detection
-from .detect_basic import (
-    _check_bottom_inputs,
-    _detect_bottom_on_selected_channel,
-    _validate_threshold,
-)
 from .freq_diff import _check_freq_diff_source_Sv, _parse_freq_diff_eq
 
 # lookup table with key string operator and value as corresponding Python operator
@@ -667,35 +663,36 @@ def frequency_differencing(
     return da
 
 
-# detect and mask seafloor
+# detect seafloor
+# Registry of supported methods
+METHODS = {
+    "basic": detect_basic,
+}
 
 
-def seafloor_mask_basic(
+def detect_seafloor(
     ds: xr.Dataset,
-    *,
-    channel: int | str,
-    var_name: str = "Sv",
-    threshold: float | tuple[float, float] = -50.0,
-    offset_m: float = 0.3,
-    surface_skip: int = 200,
+    method: str,
+    params: Dict,
 ) -> xr.DataArray:
-    """Return a boolean mask (True = above bottom) for all channels,
-    bottom detected on `channel`."""
-    Sv_sel, depth_sel, Sv_all, depth_all = _check_bottom_inputs(ds, var_name, channel)
-    tmin, tmax = _validate_threshold(threshold)
+    """
+    Detect seafloor using the selected method and return a bottom line.
 
-    bottom_depth = _detect_bottom_on_selected_channel(
-        Sv_sel, depth_sel, surface_skip=surface_skip, tmin=tmin, tmax=tmax, offset_m=offset_m
-    )
-    mask = (depth_all < bottom_depth).fillna(False).astype(bool).rename("mask")
+    Parameters
+    ----------
+    ds : xr.Dataset
+        Sv dataset with dimensions including ping_time and depth.
+    method : str
+        Name of the detection method to use (e.g., "basic", "blackwell").
+    params : dict
+        Parameters for the detection function (method-specific).
 
-    return mask.assign_attrs(
-        {
-            "mask_type": "seafloor_basic",
-            "detector_channel": str(Sv_sel.coords["channel"].item()),
-            "threshold_min": float(tmin),
-            "threshold_max": float(tmax),
-            "offset_m": float(offset_m),
-            "surface_skip_samples": int(surface_skip),
-        }
-    )
+    Returns
+    -------
+    xr.DataArray
+        1D DataArray of bottom depth per ping_time (no channel dimension).
+    """
+    if method not in METHODS:
+        raise ValueError(f"Unsupported bottom detection method: {method}")
+
+    return METHODS[method](ds, **params)
