--- echopype/mask/seafloor_detection/__init__.py
@@ -1,4 +1,0 @@
-from .api import compute_bottom
-from .utils import mask_below_seafloor
-
-__all__ = ["compute_bottom", "mask_below_seafloor"]
